fix: adjust difficulty in the mining loop
closes #154

When the timestamp is adjusted in the mining loop it is necessary to
also adjust the difficulty and difficulty threshhold.  This was not
happening before, which led to:

 1. occasional difficulty validation errors
 2. mean time-to-find-block that does not match expected mean.

This commit also adds a target_block_interval param to the mining
loop function, which enables tests to perform simulated mining at much
higher speeds, eg 1 second per block.

Changes:

 block/mod.rs:
   * Block::set_header_timestamp() -->
      Block::set_header_timestamp_and_difficulty()
   * log actual and expected valuus if difficulty validation fails
   * add test difficulty_control_matches()

 mine_loop.rs:
   * adjust difficulty and threshold in mining loop with the timestamp
   * mining_loop fn accepts a target_block_interval
   * improve test mine_ten_blocks_in_ten_seconds()