refactor: improve on/off chain notice handling
Refactors create_transaction() to enable the caller to specify which
output utxo shold use onchain global-announcement vs offchain
expected-utxo.  A method is provided to automatically determine this
based on if it appears the utxo is destined for our own wallet or not.

Also, create_transaction() now returns (unblinded) TransactionData
which includes a list of all ExpectedUtxo.  It is caller's
responsibility to inform the wallet, which mutates state.

It can be said that we now:
 * support off-chain notifications for the general case of all utxo
   going to our wallet instead of only the sub-case of change address.
 * support specifying if each utxo will use on or off-chain notification
   except for the change address, which is always off-chain.

Changes:

wallet_state:
 * add get_known_spending_keys().  placeholder to centralize logic
 * add is_wallet_utxo()

models/state/mod.rs:
 * add struct TransactionData
 * move UtxoReceiverData into utxo_receiver.rs
 * use ExpectedUtxo in place of ChangeUtxoData
 * create_transaction() returns (Transaction, TransactionData)
 * add doc-comment, example for create_transaction()
 * receiver_data param to create_transaction can now specify if each
   utxo uses Onchain or Offchain notification
 * add add_expected_utxos_to_wallet()
 * remove redundant create_transaction_with_timestamp()

rpc_server.rs:
 * add send_to_many() endpoint, which is called by send()
 * use UtxoReceiverData::auto() to determine if recipient(s) should
   be notified Onchain or Offchain
 * remove pause/resume of miner.  doesn't seem necessary.
 * simplify wallet notificaiton