refactor: sender can specify change notify method
Addresses #122

Enables caller of create_transaction() to choose whether the Change
UTXO will use OnChain or OffChain notification.

Note that caller could already do this since the last commit by
manually creating the Change UTXO so that sum(inputs) == sum(outputs).
Therefore, this change only makes it a bit easier, and is more
consistent in that the caller can easily control the notification
behavior of all outputs.

The default behavior is still OffChain notification and doc-comments
and example usage recommend this for normal usage.

Also, this fine level of control is not available in the RPC send and
send_to_many endpoints.  Only in the rust API.

Changes:
 * add change_notify_method arg to create_transaction()
 * update tests to use extra arg.
 * rename UtxoReceiverData --> UtxoReceiver
 * rename TransactionData --> TransactionDetails
 * fix: backwards logic in UtxoRecipient::auto()
 * improve doc-comment for create_transaction()
 * add ChangeNotifyMethod enum
 * add some missing doc-comments