fix(mutator set): Keep empty list up-to-date
In function `ArchivalMutatorSet::batch_remove`, the SWBFI MMR
authentication paths were being updated twice. The first time
in the call to `MutatorSetAccumulator::batch_remove`. The second
time in the call to `ArchivalMMR::batch_mutate_leaf_and_update_mps`
which is about setting the chunks to their new values. We do not
care about keeping those or any other authentication paths up to
date because that already happened in the first step.