feat(consensus): Update transaction proofs (stub)
Add a consensus program for updating a transaction proof along with
updating transaction data like the timestamp or mutator set hash.

Note that the tasm code has not been written yet. Also, the rust
source cannot be tested until we have some (transaction, single proof)
pair, which in turn depends on the tasm code for `SingleProof`.