feat(consensus): Merge transaction proofs (stub)
Add a consensus program for merging two transaction proof along with
the transaction (kernels) they pertain to.

Note that the tasm code has not been written yet. Also, the rust
source cannot be tested until we have some (transaction, single proof)
pair, which in turn depends on the tasm code for `SingleProof`.