fix(removal_records_integrity): Fix MMR auth path in Rust shadow
We forgot to assert that the leaf was authenticated in the Rust
shadowing, and when I added that check, the assertion failed! The
assertion failed since we for some reason used Merkle tree logic to
verify that authentication path, and our Merkle tree implementation caps
the tree height to 32, and for MMRs, we test for tree heights up to 64.