refactor(removal_records_integrity): Only initialize memory with needed fields
Introduce a data structure that represents the part of the witness of
the removal records integrity program that is used to initialize memory.
This gives a better separation of the non-determinism that comes from a
stream and that which is initialized in memory.

This reduces the bottle-neck table height, processor, with 0.5 % for
this program, with ~160 rows for the 2/2 benchmark.