test: add failing test for #172: ExpectedUtxo
demonstrates/tests that if wallet-db is not persisted after an
ExpectedUtxo is added, then the ExpectedUtxo will not exist after
wallet is dropped from RAM and re-created from disk.

note: this test presently FAILS, which demonstrates validity of
issue #172.

https://github.com/Neptune-Crypto/neptune-core/issues/172

A future commit/pr will fix the issue so that this test passes.