feat: Modify environment when invoking builtin `stark_verify`
Previously, the builtin would just call `triton_vm::verify` which is
sufficient for determining whether the proof is valid. What that does
not do is modify the emulated environment in sync with the actual
Triton VM environment. For instance, the `StarkVerify` snippet
consumes nondeterministic digests but `triton_vm::verify` does not.

The new code runs the snippet `StarkVerify` inside Triton VM in order
to percolate the changes made to the (native Triton VM) environment
down to the emulated environment. As a result, the right number of
nondeterministic digests will have been consumed and subsequent
builtins spawn in an in-sync environment with what the Triton VM
environment will look like when `prove` is called.