refactor(Tx Consensus): Relativize `Update` and `Merge` wrt `SingleProof`
Consensus programs `Merge` and `Update` now take the hash of the
`SingleProof` program as input and produce that same hash as output.
This indirection breaks a circular dependency: `Merge` and `Update`
now do not depend on `SingleProof`; and while `SingleProof` does
depend on `Merge` and `Update`, it can verify that, in this code
path, the arguments to `Merge` and `Update` had been proven valid
relative to itself (*i.e.*, relative to `SingleProof`).