devops: Add makefile recipe for producing expensive proofs
These expensive proofs are needed for some tests. The expensive proofs
are created in Triton VM and should be run with opt-level=3. With this
commit, these builds are stored in a new target-directory to prevent
overwrites of and from other builds.

write `make expensive-proofs` to compile for and create these proofs.
Feel free to add other tests to this recipe when appropriate.