fix(`Timelock`): Update tasm code for `Timelock`
The tasm code for Timelock was outdated. This discrepancy went
unnoticed because the tasm code was never actually tested. This
commit updates the tasm code with output from the `tasm-lang`
compiler, matching with the rust source. It also invokes the
tasm code in tests to verify that those tests halt gracefully (or
panic expectedly).