refactor: Split RR hashing from set equality check
Hashing removal records (“RRs”) is necessary for transaction validity in
both the programs “update” and “merge”. The number of removal record
lists that are to be hashed differ between those two programs.

This change allows hashing multiple lists of removal records into one
list of digests (“N-to-1”). The resulting list of digests can then be
processed further. For example, both “update” and “merge” use the
resulting list in a multiset equality check.

Co-authored-by: sword-smith <thor@neptune.cash>