feat: Use host-machine power to pick tx-proof type
The estimate of the host machine's power is used to decide what kind of
proof to share with peers: either a proof collection or a SingleProof.
Todo: Also add support for a proof that only contains the lock script
proof and maybe the type script proofs, and (probably) nothing else.

Co-authored-by: Alan Szepieniec <alan@neptune.cash>