security: Stop replay attacks
Tailor the premines' sender randomness to each network. This ensures
that the mutator set hash is unique on each network which, in turn,
ensures that transaction proofs that are valid on one network are not
valid on another.

If this was *not* done, then transactions on e.g. TestNet could be
replayed on Main net.

Co-authored-by: Alan Szepieniec <alan@neptune.cash>