test: Fix test `block_template_is_valid`
Configuration `tx_proving_capability` defaults to estimating using heuristics.
On some machines, this heuristic resolves to `ProofCollection`, in which case
transactions are produced that cannot be merged.

The fix is to override this setting using the function with more fine-grained
arguments, `create_transaction_with_prover_capability` and feed it
`TxProvingCapability::SingleProof`. This results in transactions whose validity
is supported by `SingleProof`s rather than `ProofCollection`s, which the miner
*can* merge.

When the test is run on a machine not capable of producing a `SingleProof`, it
must rely on a proof server.