mempool: Change key used to identify transactions
Add a method on transaction kernel to get a transaction ID which is a
new type that holds a digest. This digest is calculated only from fields
in the transaction that are *not* changed when the transaction is
updated to be valid under a new mutator set. In other words: This
transaction ID keeps identifying the same transaction after its witness
data has been updated to reflect the mining of a new block.