refactor: Modify fork choice rule
Prefer a new block if *it has a different height and* it has more
accumulated proof-of-work. The starred phrase is new.

Specifically, if someone announces a block with the same height,
it is ignored regardless of the accumulated proof-of-work number.

This fix destroys the incentive for miners to ignore blocks and
continue mining on the chain one block down. Specifically, if this
fix is not in place, then miners can:
 - ignore the incoming block
 - find our own block of the same height
 - win the block race because the proof-of-work of the miner's block
   happens to be larger.
This attack is particularly appealing when the miner is already
spending work producing a proof that he does not want to go to waste.