fix: Relate block hash to predecessor difficulty
Fixes an inconsistency:
 - in the mine loop, the block was determined to be valid when its
   hash was less than its own difficulty; whereas
 - the function `has_proof_of_work` relates the block's hash to the
   threshold derived from the predecessor's difficulty.

This commit fixes the discrepancy by relating the block's hash to the
predecessor's difficulty everywhere.

The main argument in favor of this choice and against the alternative
(which is relating a block's hash to its own difficulty) is the
incentive for miners to lie about their clocks, future-dating their
blocks and benefiting from a prematurely and downwards updated
difficulty.

The alternative does have an important advantage which we sacrifice.
When a significant fraction of mining power turns off or leaves the
network, the difficulty does not adjust until the next block is found,
which might take a while with the smaller hash power. With the
alternative, the effective difficulty drops the longer the network
goes without a block.