perf(peer_loop): Return batch of blocks based on 1st matching digest
Reverts a99b16a984b164675ae8764da0f4778690cd95ef. Now the batch of
blocks are returned based on the 1st match. This is done for performance
reasons, for making the common case faster. The peer is expected to
provide the list of block digests on which it can build on in a
preferred order, with the most preferred block first.