GlobalState: add test where same block is stored twice
Verify that nothing seemingly breaks if the same block is stored as tip
twice. This could, theoretically, happen with some weird race
conditions. We want to check that nothing breaks if that happens.

A better solution to allowing it, though, would maybe be if the
`GlobalState` simply ignored the block the 2nd time that it sees it.