ArchivalState: Prevent block from being written to disk twice
In case of reorganizations, the same block may be set through
`GlobalState::set_new_tip` more than once. If that happens, we don't
want to store the data again (either on disk or in the level-db). So
with this commit, we check if the block is already stored, and if it is,
we simply mark it as the new tip and nothing else.

Previously, if the block was set as tip twice, then ArchivalState would
e.g. report that it knew *two* blocks of height N even though they were
the same block. Also: both copies of this block would be considered
belonging to the canonical chain. And an assumption in the code is that
only one block at heigh N can belong to the canonical chain.

Tests are added in the next commit.