peer_loop: Check if mempool contains tx with higher quality proof
Don't inform main_loop if mempool contains tx with higher quality proof.
Also: introduce a "transfer"-variant of `Transaction` intended to be
shared with peers and uses this instead of `Transaction` as the payload
in the `Transaction` peer message enum. This ensures that transactions
backed by `PrimitiveWitness` are never sent to peers, as this would leak
secret data.