refactor: Make all proving-functions `async`
Transaction-creation that involves a lot of proving was already made
`async`. But there are other contexts in which `prove` is called, e.g.
during mining.

This ensures that all calls to `ConsensusProgram::prove` is called from
asynchronous functions.

In some test code, I had to spin up a new tokio-runtime to allow the
calling of async functions from non-async functions whose function
signature is defined in an external repo (tasm-lib). I think that's
totally fine, as long as we only do that in test-code.

This closes #208.