feat(mempool): allow caller to cap num txs for block
Add an argument to `get_transactions_for_block` such that the caller can
cap the number of transactions to include in a block. Since the miner
must merge proofs, they probably want to cap this number.

For now, the miner task does not actually use this feature. But it
should probably be definable through a CLI parameter.