fix(mempool): Remove all conflicting transactions
Previously, we only removed one transaction from the mempool if
transactions in the mempool conflicted with a new transaction. With this
commit, we remove all conflicting transaction, as the new transaction
could be the result of a merge of two transactions.

The rule that the transactions with highest fee-density stay in the
mempool is preserved.