feat(difficulty-control): Implement advance difficulty correction
Advance difficulty correction means that the difficulty for the block
currently being mined is reduced. The concrete reduction depends on
two parameters:
 - `ADVANCE_DIFFICULTY_CORRECTION_WAIT`
 - `ADVANCE_DIFFICULTY_CORRECTION_FACTOR`.
These parameters are set to 128 and 4, respectively.

If the timestamp difference between the block currently being mined
and the timestamp of the predecessor exceeds WAIT-many target block
intervals by a multiple of X, then the difficulty is automatically
reduced by FACTOR * X.

The motivation for this feature comes from avoiding long waits due to
sudden and large drops in mining power. Without this feature, the
network has to wait until it mines a block whose proof-of-work
satisfies the predecessor's difficulty. With this feature, the
effective difficulty is automatically reduced the longer the network
awaits the next block.