feat(atomic_mutex): Add synchronous `try_lock` method
Add a method on `AtomicMutex` that can return an error if the lock is
already held.

We only want *one* instance to `triton_vm::prove` to run at a time, but
for some cases, it's OK to not produce the proof right now, in case
another task is running the prover. So we plan to add a field to
`GlobalStateLock` which is a Mutex that indicates if the prover is
already running or not. For this, we want both the option of waiting
until the prover is free and of checking if it's free and otherwise
aborting.