sub: Put undetermined claims into appendix
A block now consists of four fields:
 - proof
 - appendix
 - body
 - header
The last three together are the *kernel*. The appendix consists of
undetermined claims that the block proof establishes are correct.
This list cannot live in the block body because then they affect the
hash of the block body and cannot establish anything about it.

Also: this commit introduces struct `BlockWitness`, which contains all
necessary information to produce a block. As we add more claims to the
appendix, more witnesses will need to live on `BlockWitness`.