feat: Ensure max one instance of Triton VM's prover runs at all times
With a new mutex lock in global state, we ensure that Triton VM's prover
never runs in multiple tasks simultaneously. Caller's that want to run
the prover can decide whether they want to wait for the prover to be
free, or if they want to return a failure if the prover is currently
occupied.

The drawback of this is that we have to pass lock objects around. It
adds some clutter to the code that can probably be encapsulated better
than what I have done here. It seems to be mainly a problem for tests,
though, not so much for main-line code.