test(hash_removal_record_index_sets): Fix flaky test
Fix flaky test that could separate the removal record list pointers with
too few words.

Since the pointers were randomized between 1 and 2^{64}, a value smaller
than the size of the encoding of a `Vec<RemovalRecord>` could be chosen,
making the test fail, as parts of the encoding of a Vec<RemovalRecord>
was overwritten by the next instance.

Solution: Ensure that the pointers are seperated by what I estimate is
the max size of a list of removal records (for this test).

Here's a regression test, in case anyone is interested:

```rust
    #[test]
    fn regression_test() {
        let seed: [u8; 32] = [
            0x36, 0x57, 0xa6, 0x1, 0x1c, 0xd8, 0xf7, 0x89, 0x63, 0x56, 0xc4, 0xae, 0x5d, 0xfd,
            0x26, 0x6b, 0x27, 0x73, 0x81, 0xea, 0xec, 0x6a, 0x56, 0x34, 0xfd, 0xfc, 0x78, 0x6a,
            0x10, 0xde, 0x6e, 0x59,
        ];
        let mut vm_state = HashRemovalRecordIndexSets::<5>.pseudorandom_initial_state(seed, None);
        HashRemovalRecordIndexSets::<5>.rust_shadow(&mut vm_state.stack, &mut vm_state.memory);
    }
```

Some PMD-output when running this test -- the diff for a pair of the
pointers is less than the size.
```
diff: 00000000000038448120
diff: 00000000000000024036
diff: 00000000000023788610
diff: 00000000000031265926
diff: 00000000000063885960
rrs_ptrs: 00000000000038448120, 00000000000038472156, 00000000000062260766, 00000000000093526692, 00000000000157412652
rrs.encode().len() = 54287
rrs.encode().len() = 51777
rrs.encode().len() = 51872
rrs.encode().len() = 53127
rrs.encode().len() = 53342

```