archival_state: Add function to get ms-update data to tip
Add a function that shows how to get from a previous mutator set
accumulator to the current tip, if such a path exists.

This new function returns Some(MutatorSetUpdate) iff the mutator set belongs
to the canonical chain. This functionality is needed when determining whether
to perform the `Update` consensus program on a transaction SingleProof.

If the mutator set against which the single proof is valid does not
belong to what is now the canonical chain, then we do not attempt an
upgrade.