fix(BlockProgram): Fix implementation of Block's 1st claim: transaction_is_valid
Most substantial problem was with the MAST hash verification of the
transaction kernel (mast hash) which is a leaf in the block MAST hash.
Because of the generic implementation of `merkle_tree` in the trait
`MastHash`, it's a bit roundabout how to get the leaf for the
transaction kernel in the block MAST hash: You have to call
hash_varlen(transaction_kernel_mast_hash), which is a different way of
hashing a digest than what we normally do, as the padding is different.