style: Use idiomatic `OnceLock` for caching
In `BlockPrimitiveWitness` we have all information necessary to
compute a block body but since this is a rather expensive task we want
to cache the result and use the cached version if we ever need the
body again. Previously, we used an `Option` type to achieve this, but
this choice requires passing mutable variables. `OnceLock` achieves
the same thing but without mutability because of internal mutability.

That said, fewer traits can be automatically derived for
`BlockPrimitiveWitness` because of this field type. We determined that
none of those derivations are necessary, so we dropped them.

Co-autored-by: Thorkil Schmidiger <thor@neptune.cash>