test: Fix test `mine_10_blocks_in_10_seconds`
Regrettably, preparing the block body + appendix + (placeholder-)
proof is slow: it takes approximately half a second on my desktop
machine. Mining 10 blocks in 10 seconds is possible with this
limitation, but but the parameters were set to mine 100 blocks in
10 seconds, at a rate of 1 per 100 ms. One cannot finish preparing
the block within this window, much less guess a suitable nonce.

Fixed by generalizing
 - the number of blocks to be mined, and
 - the time window for the entire test.
A few test cases call the generalized test function with hardcoded
parameters, which are compatible with success in the context of the
limitation mentioned above.