perf(test): Speedup how blocks with valid PoWs are found for tests
Prior to this commit, some tests took hours to run because the
nonce-guessing was implemented such that a whole new block-template (a
block structure) was generated for each guess. With this change, only
the nonce is updated in this test code. So finding a block with a valid
PoW is at least 100x faster.