fix(block_program): Output all claim digests verified
Output to std-out all the digests of the claims that are verified by the
program. That way, each new claim that is verified adds a new digest to
standard out. This way, it should be possible to soft-fork in future
consensus rules that apply on both the block level and on the
transaction level.