feat: add serial job-queue for triton-vm jobs
The JobQueue in src/job_queue/queue.rs is a general purpose queue for
processing heterogenous (mixed-type) jobs in a serial fashion (one after
the other) and returning a job-specific result-type back to the initiator,
who can choose to await the job or not. The queue supports job priorities
so that higher priority jobs can run before lower priority jobs.

TritonVmJobQueue is a JobQueue with a fixed/labelled set of priorities.
It is used to ensure that only one VM proving job can run at a time
because the jobs are so resource intensive (cpu and ram) that multiple
jobs run in parallel could cripple even a very powerful machine.

squashed:
* begin integration of job_queue
* TritonVmJobQueue is now just a type alias
* make GlobalStateLock::vm_job_queue private
* make a vm job for prove_consensus_program
* make handle_upgrade process tx in a loop.
* tests build
* job-queue now supports heterogenous job types and results
* fix downcast of ConsensusProgramProverJobResult
* all tests pass, hurray
* restructure job_queue module
* refactor prove_consensus_program
* make job-queue generic over priority
* pass job priority params. also refactored upgrade_job in main_loop
* revert to spawning handle_upgrade() in its own task
* fix clippy warning