ConsensusProgram: Require clean stack at halt
When a consensus program is run, require that it leaves the stack
untouched, as it was when program execution started.

That's not a requirement that is checked while verifying a proof but it
forces the consensus program developer to know what their stack looks
like, as they will otherwise pop too many or too few stack elements
prior to halt. If they inadvertently overwrite zero-initialized stack
values, that will also be caught.