chore: add log-slow-write-lock crate feature
This feature is nightly only.  To enable:
cargo +nightly build --features log-slow-write-lock

It logs a warning and caller location if the global-state
write-lock is held more than 100 millis.

Useful for finding slow writers that can harm concurrency.

It requires nightly in order to log the caller location for
async methods such as lock_guard_mut().