feat: add ScopeDurationLogger
ScopeDurationLogger is used to log a warning for scopes (eg fn) that
take longer than 1 millisecond (by default) to execute.

The duration threshold can be changed at runtime with env var:
  LOG_SLOW_SCOPE_THRESHOLD=0.000001 cargo run

The default of 1 millisecond is specified as:
  LOG_SLOW_SCOPE_THRESHOLD=0.001

The ScopeDurationLogger is used at most execution entry points, eg:
 * all rpc methods
 * peer message handlers
 * main to peer message handlers
 * peer to main message handlers

The idea is to gain a better understanding of where we are doing
slow processing.  There might be some surprises and low hanging
fruit.