fix: shutdown even if prover is running.
modifies main() to create its own tokio runtime so that we can
call tokio Runtime::shutdown_timeout().

shutdown_timeout() is used to terminate after 10 seconds even if
task(s) are still running.  Typically after 10 seconds only a
prover thread would be running.

This solution is not perfect because it is not guaranteed that all
behaved async tasks would have terminated already, eg one could be
waiting on DB to flush files.

However it is a simple solution that improves on the previous
behavior which prevented shutdown from finishing if the prover is
running.  Repeated ctrl-c did not work.  The only way to exit was
to kill -9 the process.

With this fix, ctrl-c works even if the prover is running.  Though
it will take up to 10 seconds to actually exit.