feat: Two-Step Mining
Mining is now divided into two seperate tasks that where the user can
choose which to participate in. The tasks are: composing and guessing.

Composing refers to creating valid blocks where creating the associated
STARK proofs is the expensive part. The result of a composition is
referred to as a block proposal. Guessing refers to using a block
proposal from the composition-step in order to attempt to find a valid
PoW solution through the process of hashing.

If a guesser is not also composing, they can get the block proposals
from peers.

Specifically, this commit changes / adds:

 - New type of `BlockProposal` and field of this type on `GlobalState`
 a) in peer-to-peer communication we need to know whether the current
    block proposal is our own or not (in order to make sure not to
    guess on top of a foreign proposal),
 b) we need to keep track of information specific to our own proposal,
    namely the expected UTXOs claimable by the composer.
 - Update CLI flags to specify guessing and composing separately.
 - Define function `mine` on CLI arguments that determines whether the
   mine loop should be started, as a function of the `guess` and
   `compose` CLI flags.
 - Communication:
   - Peers exchange block proposal notifications and block proposals;
   - Miner informs Main of own block proposals;
   - Main informs Miner of foreign block proposals (if applicable);
   - Main informs Peers of block proposals.
 - Anticipate multiple Expected UTXOs for composer

Co-authored-by: Alan Szepieniec <alan@neptune.cash>