ArchivalMutatorSet: Fix get_mutator_set_update_to_tip (#227)
Fix a problem in get_mutator_set_update_to_tip where the wrong removal
records were returned.

The function attempts to find a `MutatorSetUpdate` that transforms the
mutator set from a state B to a later state A where B and A can be
separated by an arbitrary number of blocks. The problem was the the
removal records were collecting from all blocks between B and A. But the
MutatorSetUpdate data structure needs the removal records valid for
block B.

To fix this, we make ephemeral changes to the archival mutator set to
assemble the required information. And then roll those ephemeral changes
back once the information is gathered.

This commit includes a test that the function `get_mutator_set_update_to_tip`
does not mutate the archival mutator set.

This closes #225.