test: add spawned_tasks_live_as_long_as_jobqueue
This attempts to test that tasks spawned by JobQueue
live until JobQueue is dropped, even while tokio executor
is shutting down.

The test is successful in that it replicates a problem
we observed in neptune-core, but so far no solution has been
found so the test fails.

It is presently marked with #[should_panic] to prevent CI
problems until soluiton is found or test removed.