feat: Reward peers for good behavior
This commit adds a `reward` function on `PeerLoopHandler` that
improves peers' standings when they
 - send valid blocks
 - send a valid block proposal
 - send a valid transaction.

Notifications should not be rewarded because the thing they inform
the recipient of might not be valid.

This commit also refactors `PeerInfo` and `PeerStanding` to better
encapsulate logic and data and present the outside world with a
smaller, abstract interface.