mempool: Fix wrong answer about proof quality
Return false in the case where mempool knows the transaction but the
transaction is *not* backed by an actual STARK proof but rather a
primitive witness that cannot be safely shared with peers. In this
(unlikely) situation, we should just request the transaction from a peer
if they offer it.