mempool: Track which transaction are own
This is done to determine which transactions that should be updated when
a new block arrives.

Update is an expensive STARK proof to produce, so we opt for the
following policy:
- own transactions are updated, if possible.
- If client is composing, all transactions are updated, as we otherwise
  cannot include the mempool transactions in the new block.

Also note that, if client is composing, it *has* to wait for the update
to finish before starting a new composition. Otherwise, the transactions
will never be mined.