feat: Add CLI flag `--no-transaction-initiation`
This flag, when set, prevents the node from initiating a transaction.
Specifically, the RPC calls for initiating transactions (`send` and
`send_to_many`) abort early if this flag is set.

Also: modify interface of `mock_genesis_global_state`, which now takes
an extra argument: the CLI arguments. This change originates from the
unit test which requires the relevant flag to be set in the CLI
arguments field of global state.