refactor(`Block`): Reduce visibility of `mutator_set_accumulator`
The field `mutator_set_accumulator` lives on `BlockBody`, not `Block`.
This commit reduces the visibility of `mutator_set_accumulator` and
introduces an accessor living on `Block`, which is used instead of
accessing the field directly. This change anticipates a more profound
one in which the meaning of this field changes: whereas up until now
it defines the state of the mutator set in between two blocks; in
future it will represent an in-between state, *after* applying the
update induced by the transaction but *before* applying the update
induced by the guesser fees (and perhaps in the further future, that
induced by composer fees).