mempool: spawn threads for proof-updates
Change behavior of `Mempool::update_with_block_and_predecessor` to spawn
threads for the update STARK proof production instead of doing the
update within the execution of this method. When this method is called
through `GlobalState`'s `set_new_tip` a write-lock is held over the
global state. So producing the STARK proofs in a blocking manner would
hold this write lock for minutes, which is obviously unacceptable.

This funciton returns the join handles to the caller such that callbacks
can be invoked.