fix: ignore stderr when invoking triton-vm-prover
triton-vm-prover sometimes panics and emits a backtrace when it is
killed, eg when neptune-core is shutting down, or a proving job is
cancelled for any reason.  The panic message was being displayed
in the console because triton-vm-prover inherited stderr from the
parent process.

for now we simply pipe stderr to null (/dev/null) so it is not
displayed.

eventually triton-vm-prover should be made more robust/graceful
when it receives a kill signal, but this is adequate for now.