chore: change how queued jobs are counted
previously when a job begins processing the jobqueue would log a count
of jobs in the queue, including the newly processing job.

as such, the count would never reach 0.

we no longer the count the newly processing job, as it has already
been removed from the queue.  it seems to make things clearer when
reading log output.