refactor: Distinguish between block-body-MSA and MSA-after-block
Specifically: the mutator set accumulator in the block body is *not*
the mutator set accumulator after the block is applied. The difference
is that the block defines miner fee UTXOs (right now only guesser fee
UTXOs) which are added to it also.

Function `Block::mutator_set_accumulator_after()` is renamed (adding
`_after`) to clarify.

This commit adds arbitrary implementations allowing for the generation
of an arbitrary block primitive witness. This arbitrary implementation
reflects the new dependency graph.

Where possible, apply or un-apply mutator set update derived from
block. Use the interface that hides the internal mechanics.