fix: stop queue when last JobQueue instance drops
Previously the job queue was being stopped when any instance of the
JobQueue is dropped.

Since JobQueue resides in GlobalStateLock and it gets cloned into
spawned tasks then any time a task would end (such as mining) then
a JobQueue would be dropped, which would stop the job queue.

The fix is to wait until the last JobQueue instance is dropped
before stopping the queue.

In fixing this, I also discovered that jobs were not correctly
being cancelled when the queue is stopped.  A fix for that is
also included.

a new test stop_only_called_once_when_cloned verifies both
fixes.