main_loop: Handle mempool tx-updating in spawned task
Instead of runinng the prover udring the call to set_new_tip, this
commit spawns a tokio task for that task. This means that a write lock
over the global state is *not* held during the production of the
expensive STARK proofs.

This spawned task is responsible for informing the miner that it can
start on the new block, once the task is done updating the proofs.