devx: Make failed VM run error readable
Prior to logging a proof job, the computation is run on the VM to
verify graceful halt. If this program (on the given input, etc.)
does not halt gracefully, then the VM returns an error object of
type `VMError`. The VMError type implements `Display` and its
rendering to the screen is helpful for debugging.

This commit prints the helpful debugging information, which was
previously hidden behind error-ignoring asserts.