TransactionProof: Delete proof type invalid
When removing this type, TransactionProof no longer had a Default
implementation which meant (for some reason) it coulnd't be hashed. This
made me discover that the keys that wallet uses for unconfirmed
trnansaction were of the wrong type. We should use the same key type as
in the mempool, namely TransactionKernelId.