feat: cli opt max_log2_padded_height_for_proofs
this cli option enables user to limit the max complexity
of proofs.

A proving job will return an error if it encounters a program
with a proof over the limit.

By default no limit is set.

At present when the job returns an error calling code up the
stack will panic, and the panic is swallowed by tokio.

A future commit will address these issues.

Of equal benefit, this commit sets up a framework for passing
settings to the proving job. Later we can add new setting fields
without need to modify all the intermediate functions.