chore: use panic=abort for release and dev builds
This fixes cases where tokio runtime would catch panics from
detached spawned tasks, log them, and continue on.

This is bad because it is easy for the panic to go un-noticed in
the log.  A task such as mining or job-queue may have stopped running
but neptune-core continues on and no problem is apparent.

panic=abort causes the panic!() call to abort immediately rather
than attempting to unwind the stack and cleanup.  As such, it is
a simple change that "fixes" all such cases.

other options include:
 * call catch_unwind() on futures before spawning them.
 * await/select join-handle for each spawned task so that it
   is no longer detached.  if it panics or returns an error it will
   be propagated up.

these options may be more elegant but are also more complex to
implement correctly.