refactor: Maybe future-date Tx timestamp in block proposal
Problem: shared block proposals were rejected as invalid because the
timestamp is less than `MINIMUM_BLOCK_TIME` ahead of the predecessor.

Explanation: the timestamp of the block proposal denotes the time that
proving started. Proving starts immediately after the previous block
was found, so `MINIMUM_BLOCK_TIME` has not passed.

Solution: set the timestamp of the block proposal to now or to the
timestamp of the predecessor block plus the `MINIMUM_BLOCK_TIME`,
whichever is later.

Also: add test that verifies that a block proposal for height 2 is
valid.

Co-authored-by: Thorkil Schmidiger <thor@neptune.cash>