mine_loop: Inject timestamp to `compose_block`
The current timestamp is injected to this function to allow us to test
this function in the test `block_proposal_for_height_two_is_valid`. This
test acts as a regression test for issue #233^[0] which identified a
problem with invalid block proposals. These invalid block proposals were
rejected by peers because their timestamp was too close to the
predecessor block. A line in `compose_block` now ensures that the
timestamp is at least MINIMUM_BLOCK_TIME larger than the previous
block's timestamp. That (the `max(now, parent + min)`) was introduced in
ed9df050e73809b504e37318adf74e1be0062265 but the test introduced there
did not work. This commit fixes that test.

[0]: https://github.com/Neptune-Crypto/neptune-core/issues/233