feat(dashboard): Add mempool overview screen
This commit adds a new screen to the dashboard, giving an overview of
the transactions in the mempool.

To support this screen, the RPC adds a call, `mempool_overview` which
takes a start index and page size, and
returns a `Vec` of `MempoolTransactionInfo` objects (also a new
struct). Besides the obvious, these objects also contain info about:
 - how they affect the balance (wallet state got a new public
   function to support this)
 - whether the transaction is synced.
To compute this list, the RPC call handler computes a mildly expensive
join operation between data from various sources.

Co-authored-by: Thorkil Schmidiger