test: Fix mempool-related test in CI
The prover capability must be set prior to the creation of the global
state, since the prover capability is read from network state and not
from CLI.

The fundamental problem is that the CI machines are single-cored, so
they'll be set with very weak proving capability. We must force it to
the highest level for this test to work, and that forcing must happen
before the global state is created.